recodeGADS: Fix sequential recoding for NAs
Treats NAs consistently as all other oldValues in recodeGADS. eatTools::recodeLookup() can deal with NAs, therefore NAs are simply handed through to it.

Duplicate values in oldValues are now tested for directly in recodeGADS(), to provide an informative error message if this is not forwarded to applyChangeMeta().

In the tests for applyChangeMeta(), inconsistently recodeGADS() was used. This has been fixed as well.